adtelligent bid adapter: add new allias (#12559)
* addStellormedia alias

* add adtellintUtils

* update adtelligentUtils

* fix test

---------
