+++ test/spec/modules/adtelligentBidAdapter_spec.js
@@ -17,6 +17,7 @@ const aliasEP = {
   'ocm': 'https://ghb.cenarius.orangeclickmedia.com/v2/auction/',
   '9dotsmedia': 'https://ghb.platform.audiodots.com/v2/auction/',
   'indicue': 'https://ghb.console.indicue.com/v2/auction/',
+  'stellormedia': 'https://ghb.ads.stellormedia.com/v2/auction/',
 };
 
 const DEFAULT_ADATPER_REQ = { bidderCode: 'adtelligent' };
